Remove dependency on ppx_bin_prot and ppx_deriving_yojson
The values provided by those derivers were deprecated anyway.  The logic
behind those deprecations was that it's not key-parsers' job to define
and implement serialization formats for keys.  Since serialization isn't
trivial (the format must be stable or there must be a migration path
from one format to the next), it's better left to the user to define
whatever they need for their application.