Handle negative RSA keys
A few tools omit to add a leading 0-bit at the start of their keys,
leading to the integer being read as negative. The operation to fix
this would be a 2's complement, but as it is hard to know the length
of the underlying bitfield, the module will just return an error
instead of trying to fix the key.